requested changes, + sugg
I have applied the requested changes, except for removing 1 of the resources.
I would like to suggest that Free Pik is still relevant. It contains downloadable fonts, and other vectors/graphics that are usable when creating html/css (backgrounds, clickable objects, etc.). It also has templates for website mockups/wireframe type building.
I would classify it as a similar resource to that of Canva wireframe templates, and/or awesomefonts.

Let me know the final determination and I can remove if needed.